+++ lib/common/interfaces/local_preferences_service.dart
@@ -1,0 +1,4 @@
+abstract class ILocalPreferencesService {
+  Future<T?> read<T>({required String key});
+  Future<bool> write<T>({required String key, required T value});
+}
