--- lib/core/app/data/services/local_storage_service.dart
@@ -1,5 +1,0 @@
-import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/common/services/shared_preferences_service.dart';
-
-@lazySingleton
-class LocalStorageService extends SharedPreferencesService {}
