+++ lib/core/app/domain/services/manage_files_service.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/core/services/app_service.dart';
-import 'package:tv_randshow/core/services/databases/i_database_service.dart';
+import 'package:tv_randshow/core/app/data/services/app_service.dart';
+import 'package:tv_randshow/core/tvshow/domain/interfaces/i_database_service.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/file.dart';
 
 @lazySingleton
