+++ lib/core/migration/domain/use_cases/add_streamings_migration_use_case.dart
@@ -2,10 +2,10 @@ import 'dart:ui' as ui;
 
 import 'package:injectable/injectable.dart';
 import 'package:tv_randshow/core/migration/domain/models/migration_status.dart';
-import 'package:tv_randshow/core/services/databases/i_database_service.dart';
 import 'package:tv_randshow/core/streaming/domain/interfaces/i_streamings_repository.dart';
 import 'package:tv_randshow/core/streaming/domain/models/streaming.dart';
 import 'package:tv_randshow/core/streaming/domain/models/streaming_search.dart';
+import 'package:tv_randshow/core/tvshow/domain/interfaces/i_database_service.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 @injectable
