+++ lib/core/migration/domain/use_cases/verify_old_database_use_case.dart
@@ -1,5 +1,5 @@
 import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/core/services/databases/i_secondary_database_service.dart';
+import 'package:tv_randshow/core/tvshow/data/repositories/i_secondary_database_service.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 @Injectable(env: ['mobile'])
