+++ lib/core/tvshow/domain/interfaces/i_database_service.dart
