+++ lib/ui/viewmodels/widgets/details_model.dart
