+++ lib/ui/viewmodels/widgets/favorite_list_model.dart
@@ -1,7 +1,7 @@
 import 'package:stacked/stacked.dart';
 import 'package:tv_randshow/config/locator.dart';
-import 'package:tv_randshow/core/services/app_service.dart';
-import 'package:tv_randshow/core/services/favs_service.dart';
+import 'package:tv_randshow/core/app/data/services/app_service.dart';
+import 'package:tv_randshow/core/app/domain/services/favs_service.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/tvshow_actions.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
