+++ lib/ui/widgets/save_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_translate/flutter_translate.dart';
 import 'package:stacked/stacked.dart';
-import 'package:tv_randshow/core/viewmodels/widgets/save_model.dart';
+import 'package:tv_randshow/ui/viewmodels/widgets/save_model.dart';
 import 'package:unicons/unicons.dart';
 
 class SaveButton extends StatelessWidget {
